Move Direction3D to avoid circular import
